Move DiscussionCourseXBlock to openedx/core/lib/xblock_builtin/xblock_discussion/xblock_discussion.py.
