+++ lms/templates/discussion/_discussion_course.html
@@ -3,11 +3,11 @@
 <%namespace name='static' file='../static_content.html'/>
 
 <%block name="js_extra">
-    <%static:js group='discussion' raw='true'/>
+    <%static:js group='discussion'/>
 </%block>
 
 <%block name="css_extra">
-    <%static:css group='discussion' raw='true'/>
+    <%static:css group='discussion'/>
 </%block>
 
 <div class="discussion-course">
