+++ openedx/core/lib/xblock_builtin/xblock_discussion/__init__.py
@@ -1,0 +1,1 @@
+from .xblock_discussion import DiscussionXBlock, DiscussionCourseXBlock
