+++ openedx/core/lib/xblock_builtin/xblock_discussion/utils.py
@@ -1,0 +1,8 @@
+from django.templatetags.static import static
+
+def asset_to_static_url(asset_path):
+    """
+    :param str asset_path: path to asset
+    :return: str|unicode url of asset
+    """
+    return static(asset_path)
